create shardingsphere-kernal module
