+++ shardingsphere-kernal/shardingsphere-kernal-context/src/main/java/org/apache/shardingsphere/kernal/context/SchemaContexts.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.context;
+package org.apache.shardingsphere.kernal.context;
 
 import lombok.Getter;
 import org.apache.shardingsphere.infra.auth.Authentication;
